quiqueypatrysecasan/requirements.txt
Django==1.9.5
django-crispy-forms==1.6.0
Pillow==3.2.0
gunicorn
psycopg2
cloudinary